+++ zython/operations/functions_and_predicates.py
@@ -3,19 +3,17 @@ from typing import Union, Callable, Optional
 import zython
 from zython.operations import _iternal
 from zython.operations import constraint as constraint_module
-from zython.operations import operation as operation_module
 from zython.operations._op_codes import _Op_code
 from zython.operations.constraint import Constraint
 from zython.operations.operation import Operation
-from zython.var_par.collections.array import ArrayMixin
 from zython.var_par.collections.set import SetVar
+from zython.var_par.get_type import get_base_type, derive_operation_type
 from zython.var_par.types import ZnSequence
 from zython.var_par.var import var
 
 
-def exists(seq: ZnSequence,
-           func: Optional[Union["Constraint", Callable]] = None) -> Constraint:
-    """ Specify constraint which should be true for `at least` one element in ``seq``.
+def exists(seq: ZnSequence, func: Optional[Union["Constraint", Callable]] = None) -> Constraint:
+    """Specify constraint which should be true for `at least` one element in ``seq``.
 
     The method has the same signature as ``forall``.
 
@@ -42,8 +40,7 @@ def exists(seq: ZnSequence,
     return constraint_module._exists(seq, iter_var, operation)
 
 
-def forall(seq: ZnSequence,
-           func: Optional[Union["Constraint", Callable]] = None) -> Constraint:
+def forall(seq: ZnSequence, func: Optional[Union["Constraint", Callable]] = None) -> Constraint:
     """
     Takes expression (that is, constraint) or function which return constraint
         and make them a single constraint which should be true for every element in the array.
@@ -77,9 +74,8 @@ def forall(seq: ZnSequence,
     return constraint_module._forall(seq, iter_var, operation)
 
 
-def sum(seq: ZnSequence,
-        func: Optional[Union["Constraint", Callable]] = None) -> Operation:
-    """ Calculate the sum of the ``seq`` according with ``func``
+def sum(seq: ZnSequence, func: Optional[Union["Constraint", Callable]] = None) -> Operation:
+    """Calculate the sum of the ``seq`` according with ``func``
 
     Iterates through elements in seq and calculate their sum, you can modify summarized expressions
     by specifying ``func`` parameter.
@@ -89,8 +85,8 @@ def sum(seq: ZnSequence,
     seq: range, array of var, or sequence (list or tuple) of var
         sequence to sum up
     func: Operation or Callable, optional
-        Operation which will be executed with every element and later sum up. Or function which returns
-        such operation.
+        Operation which will be executed with every element and later sum up. 
+        Or function which returns such operation.
         If function or lambda it should be with 0 or 1 arguments only.
 
     Returns
@@ -124,17 +120,60 @@ def sum(seq: ZnSequence,
     Solution(objective=5, a=4, b=3, c=5)
     """
     iter_var, operation = _iternal.get_iter_var_and_op(seq, func)
-    if isinstance(seq, ArrayMixin) and operation is None:
-        type_ = seq.type
-    else:
-        type_ = operation.type
-    if type_ is None:
-        raise ValueError("Can't derive the type of {} expression".format(func))
-    return operation_module._sum(seq, iter_var, operation, type_=type_)
+    type_ = derive_operation_type(seq, operation)
+    return Operation(_Op_code.sum_, seq, iter_var, operation, type_=type_)
+
+def product(seq: ZnSequence, func: Optional[Union["Constraint", Callable]] = None) -> Operation:
+    """Calculate the product of the ``seq`` according to ``func``
+
+    Iterates through elements in seq and calculates their product, you can modify the expressions
+    being multiplied by specifying the ``func`` parameter.
+
+    Parameters
+    ----------
+    seq: range, array of var, or sequence (list or tuple) of var
+        sequence to multiply
+    func: Operation or Callable, optional
+        Operation which will be executed with every element and later multiplied.
+        Or function which returns such operation.
+        If function or lambda it should be with 0 or 1 arguments only.
+
+    Returns
+    -------
+    result: Operation
+        Operation which will calculate the product
+
+    Examples
+    --------
+
+    >>> import zython as zn
+    >>> class MyModel(zn.Model):
+    ...     def __init__(self):
+    ...         self.a = zn.Array(zn.var(range(1, 5)), shape=4)
+    >>> model = MyModel()
+    >>> model.solve_minimize(zn.product(model.a))
+    Solution(objective=1, a=[1, 1, 1, 1])
+
+    # find the product of squares of elements
+
+    >>> import zython as zn
+    >>> class MyModel(zn.Model):
+    ...     def __init__(self):
+    ...         self.a = zn.var(zn.range(4))
+    ...         self.b = zn.var(zn.range(4))
+    ...         self.c = zn.var(zn.range(4))
+    ...         self.constraints = [zn.product((self.a, self.b, self.c), lambda i: i ** 2) == 36]
+    >>> model = MyModel()
+    >>> model.solve_satisfy()
+    Solution(a=3, b=2, c=1)
+    """
+    iter_var, operation = _iternal.get_iter_var_and_op(seq, func)
+    type_ = derive_operation_type(seq, operation)
+    return Operation(_Op_code.product, seq, iter_var, operation, type_=type_)
 
 
 def count(seq: ZnSequence, value: Union[int, Operation, Callable[[ZnSequence], Operation]]) -> Operation:
-    """ Returns the number of occurrences of ``value`` in ``seq``.
+    """Returns the number of occurrences of ``value`` in ``seq``.
 
     Parameters
     ----------
@@ -183,16 +222,16 @@ def count(seq: ZnSequence, value: Union[int, Operation, Callable[[ZnSequence], O
     Counter({3: 1, 2: 1, 1: 1, 0: 1})
     """
     iter_var, operation = _iternal.get_iter_var_and_op(seq, value)
-    return operation_module._count(seq, iter_var, operation, type_=int)
+    return Operation(_Op_code.count, seq, iter_var, operation, type_=int)
 
 
 def cumulative(
-        start_times: ZnSequence,
-        durations: ZnSequence,
-        requirements: ZnSequence,
-        limit: Union[int, var],
+    start_times: ZnSequence,
+    durations: ZnSequence,
+    requirements: ZnSequence,
+    limit: Union[int, var],
 ) -> Constraint:
-    """ The cumulative constraint is used for describing cumulative resource usage.
+    """The cumulative constraint is used for describing cumulative resource usage.
 
     It requires that a set of tasks given by start times, durations, and resource requirements,
     never require more than a global resource limit at any one time.
@@ -240,15 +279,15 @@ def cumulative(
     >>> result["limit"]
     2
     """
-    return constraint_module.cumulative(start_times, durations, requirements, limit)
+    return Constraint(_Op_code.cumulative, start_times, durations, requirements, limit)
 
 
 def disjunctive(
-        start_times: ZnSequence,
-        durations: ZnSequence,
-        strict: bool = False,
+    start_times: ZnSequence,
+    durations: ZnSequence,
+    strict: bool = False,
 ) -> Constraint:
-    """ The disjunctive constraint takes an array of start times for each task and
+    """The disjunctive constraint takes an array of start times for each task and
     an array of their durations and makes sure that only one task is active at any one time.
 
     Parameters
@@ -324,17 +363,12 @@ def table(
     >>> result["a"]
     [1, 2, 3, 4]
     """
-    return constraint_module.table(x, t)
+    return Constraint(_Op_code.table, x, t)
 
 
 def abs(x: float | var) -> Operation:
-    """ Return absolute value of x
+    """Return absolute value of x
 
-    Parameters
-    ----------
-    x: float, var or par
-        value to get absolute value of
-        
     Returns
     -------
     result: Operation
@@ -356,11 +390,149 @@ def abs(x: float | var) -> Operation:
     >>> model.solve_satisfy()
     Solution(a=1, v=3.0, b=1, c=3.0)
     """
-    return operation_module._abs(x)
+    return Operation(_Op_code.abs, x, type_=get_base_type(x))
+
+def exp(x: float | var) -> Operation:
+    """Return the exponential of x
+
+    Returns
+    -------
+    result: Operation
+        Operation which will calculate the exponential.
+
+    Examples
+    --------
+
+    >>> import zython as zn
+    >>> class MyModel(zn.Model):
+    ...     def __init__(self):
+    ...         self.a = zn.exp(1)
+    ...         self.b = zn.var(float)
+    ...         self.constraints = [self.b == zn.exp(2)]
+    >>> model = MyModel()
+    >>> model.solve_satisfy()
+    Solution(a=2.718281828459045, b=7.38905609893065)
+    """
+    return Operation(_Op_code.exp, x, type_=float)
+
+def ln(x: float | var) -> Operation:
+    """Return the natural logarithm of x
+
+    Returns
+    -------
+    result: Operation
+        Operation which will calculate the natural logarithm.
+
+    Examples
+    --------
+
+    >>> import zython as zn
+    >>> class MyModel(zn.Model):
+    ...     def __init__(self):
+    ...         self.a = zn.ln(7.38905609893065)
+    ...         self.b = zn.var(float)
+    ...         self.constraints = [self.b == zn.ln(2.718281828459045)]
+    >>> model = MyModel()
+    >>> model.solve_satisfy()
+    Solution(a=2.0, b=1.0)
+    """
+    return Operation(_Op_code.ln, x, type_=float)
+
+def log(x: float | var, base: float) -> Operation:
+    """Return the logarithm of x to the given base
+
+    Returns
+    -------
+    result: Operation
+        Operation which will calculate the logarithm.
+
+    Examples
+    --------
+
+    >>> import zython as zn
+    >>> class MyModel(zn.Model):
+    ...     def __init__(self):
+    ...         self.a = zn.log(8, 2)
+    ...         self.b = zn.var(float)
+    ...         self.constraints = [self.b == zn.log(27, 3)]
+    >>> model = MyModel()
+    >>> model.solve_satisfy()
+    Solution(a=3.0, b=3.0)
+    """
+    return Operation(_Op_code.log, base, x, type_=float)
+
+def log10(x: float | var) -> Operation:
+    """Return the base-10 logarithm of x
+
+    Returns
+    -------
+    result: Operation
+        Operation which will calculate the base-10 logarithm.
+
+    Examples
+    --------
+
+    >>> import zython as zn
+    >>> class MyModel(zn.Model):
+    ...     def __init__(self):
+    ...         self.a = zn.log10(100)
+    ...         self.b = zn.var(float)
+    ...         self.constraints = [self.b == zn.log10(1000)]
+    >>> model = MyModel()
+    >>> model.solve_satisfy()
+    Solution(a=2.0, b=3.0)
+    """
+    return Operation(_Op_code.log10, x, type_=float)
+
+def log2(x: float | var) -> Operation:
+    """Return the base-2 logarithm of x
+
+    Returns
+    -------
+    result: Operation
+        Operation which will calculate the base-2 logarithm.
+
+    Examples
+    --------
+
+    >>> import zython as zn
+    >>> class MyModel(zn.Model):
+    ...     def __init__(self):
+    ...         self.a = zn.log2(8)
+    ...         self.b = zn.var(float)
+    ...         self.constraints = [self.b == zn.log2(16)]
+    >>> model = MyModel()
+    >>> model.solve_satisfy()
+    Solution(a=3.0, b=4.0)
+    """
+    return Operation(_Op_code.log2, x, type_=float)
+
+def sqrt(x: float | var) -> Operation:
+    """Return the square root of x
+
+    Returns
+    -------
+    result: Operation
+        Operation which will calculate the square root.
+
+    Examples
+    --------
+
+    >>> import zython as zn
+    >>> class MyModel(zn.Model):
+    ...     def __init__(self):
+    ...         self.a = zn.sqrt(4)
+    ...         self.b = zn.var(float)
+    ...         self.constraints = [self.b == zn.sqrt(9)]
+    >>> model = MyModel()
+    >>> model.solve_satisfy()
+    Solution(a=2.0, b=3.0)
+    """
+    return Operation(_Op_code.sqrt, x, type_=float)
 
 
 def min(seq: ZnSequence, key: Union[Operation, Callable[[ZnSequence], Operation], None] = None) -> Operation:
-    """ Finds the smallest object in ``seq``, according to ``key``
+    """Finds the smallest object in ``seq``, according to ``key``
 
     Parameters
     ----------
@@ -391,11 +563,12 @@ def min(seq: ZnSequence, key: Union[Operation, Callable[[ZnSequence], Operation]
     Solution(m=-3)
     """
     iter_var, operation = _iternal.get_iter_var_and_op(seq, key)
-    return operation_module._min(seq, iter_var, operation, type_=int)
+    type_ = derive_operation_type(seq, operation)
+    return Operation(_Op_code.min_, seq, iter_var, operation, type_=type_)
 
 
 def max(seq: ZnSequence, key: Union[Operation, Callable[[ZnSequence], Operation], None] = None) -> Operation:
-    """ Finds the biggest object in ``seq``, according to ``key``
+    """Finds the biggest object in ``seq``, according to ``key``
 
     Parameters
     ----------
@@ -429,11 +602,12 @@ def max(seq: ZnSequence, key: Union[Operation, Callable[[ZnSequence], Operation]
     Solution(m=3)
     """
     iter_var, operation = _iternal.get_iter_var_and_op(seq, key)
-    return operation_module._max(seq, iter_var, operation, type_=int)
+    type_ = derive_operation_type(seq, operation)
+    return Operation(_Op_code.max_, seq, iter_var, operation, type_=type_)
 
 
 class alldifferent(Constraint):
-    """ requires all the variables appearing in its argument to be different
+    """requires all the variables appearing in its argument to be different
 
     Parameters
     ----------
@@ -490,11 +664,12 @@ class alldifferent(Constraint):
     >>> Counter(result["a"]) == {3: 1, 2: 1, 1: 2}
     True
     """
+
     def __init__(
-            self,
-            seq: ZnSequence,
-            except0: Optional[bool] = None,
-            except_: Union[set, zython.var_par.collections.set.Set, None] = None,
+        self,
+        seq: ZnSequence,
+        except0: Optional[bool] = None,
+        except_: Union[set, zython.var_par.collections.set.Set, None] = None,
     ):
         if all((except0, except_)):
             raise ValueError("Arguments `except0` and `except_` can't be set at the same time")
@@ -509,7 +684,7 @@ class alldifferent(Constraint):
 
 
 class allequal(Constraint):
-    """ requires all the variables appearing in its argument to be equal
+    """requires all the variables appearing in its argument to be equal
 
     Parameters
     ----------
@@ -533,12 +708,13 @@ class allequal(Constraint):
     >>> model.solve_satisfy()
     Solution(a=[[5, 5, 5, 5], [5, 5, 5, 5]])
     """
+
     def __init__(self, seq: ZnSequence):
         super().__init__(_Op_code.allequal, seq)
 
 
 class ndistinct(Operation):
-    """ returns the number of distinct values in ``seq``.
+    """returns the number of distinct values in ``seq``.
 
     Parameters
     ----------
@@ -568,12 +744,13 @@ class ndistinct(Operation):
     >>> len(set(result["a"]))
     3
     """
+
     def __init__(self, seq: ZnSequence):
         super().__init__(_Op_code.ndistinct, seq)
 
 
 class circuit(Constraint):
-    """ Constrains the elements of ``seq`` to define a circuit where x[i] = j means that j is the successor of i.
+    """Constrains the elements of ``seq`` to define a circuit where x[i] = j means that j is the successor of i.
 
     Examples
     --------
@@ -587,12 +764,13 @@ class circuit(Constraint):
     >>> model.solve_satisfy()
     Solution(a=[2, 4, 3, 1, 0])
     """
+
     def __init__(self, seq: ZnSequence):
         super().__init__(_Op_code.circuit, seq)
 
 
 def increasing(seq: ZnSequence, *, allow_duplicate: Optional[bool] = True) -> Constraint:
-    """ Requires that the sequence `seq` is in increasing order
+    """Requires that the sequence `seq` is in increasing order
 
     Parameters
     ----------
@@ -626,7 +804,7 @@ def increasing(seq: ZnSequence, *, allow_duplicate: Optional[bool] = True) -> Co
 
 
 def decreasing(seq: ZnSequence, *, allow_duplicate: Optional[bool] = True) -> Constraint:
-    """ Requires that the sequence `seq` is in decreasing order
+    """Requires that the sequence `seq` is in decreasing order
 
     Parameters
     ----------